requirements.txt
playwright
google-generativeai
chromadb
beautifulsoup4